modify controller division
