+++ src/main/java/tourGuide/controller/AttractionController.java
@@ -34,4 +34,5 @@ public class AttractionController {
         return JsonStream.serialize(tourGuideService.getNearByAttractions(visitedLocation));
     }
 
+    //TODO : obtenir les infos sur les attractions
 }
